chore(quill-charts): extract bar drawing into utils/draw-bar-chart
Move the canvas-drawing passes out of `BarChart.tsx` into a co-located
`utils/draw-bar-chart.ts`:

- `drawBarChartStatic` — the full static pass (grid, bars, tracks, shadow, pills).
- `resolveBarHoverItems` / `drawBarHoverItems` — hover resolution split from the
  paint loop. The component keeps only the `useRef` fade bookkeeping (the one
  genuinely stateful piece), owns the fade alpha, and hands the resolved result
  to the draw step.
- `stackPillRects` (and its `ALL_CORNERS` constant) — shared band-pill geometry.

`BarChart.tsx` is now a thin orchestration layer (~750 → 374 lines): config
destructuring, the memoized scale builder, and thin `useCallback` wrappers that
delegate to the extracted functions — matching `LineChart`'s altitude.

Pure refactor — no behavior or public API change. The extracted functions are
byte-for-byte the same logic with parameters threaded through explicitly, so the
existing DOM-level `BarChart`/`TimeSeriesBarChart` tests exercise every drawing
path unchanged. Final of a few small PRs breaking up the oversized component.
Stacks on `posthog-code/quill-barchart-3-click-routing`.

Generated-By: PostHog Code
Task-Id: 8e22d6ce-b20c-4948-8774-4b148bce553b
