+++ packages/quill/packages/charts/src/charts/BarChart/BarChart.tsx
@@ -1,27 +1,8 @@
-import { color as d3Color } from 'd3-color'
 import React, { useCallback, useMemo, useRef } from 'react'
 
-import {
-    bandCenter,
-    type BarChartPrivate,
-    computeBarTrackRect,
-    computeSeriesBars,
-    groupedBarCenter,
-} from '../../core/bar-layout'
-import {
-    BAR_TRACK_HOVER_ALPHA,
-    type BarRect,
-    type BarRoundedCorners,
-    clipToRoundedRects,
-    drawBarHighlight,
-    drawBars,
-    drawBarTracks,
-    drawGrid,
-    type DrawContext,
-} from '../../core/canvas-renderer'
+import { bandCenter, type BarChartPrivate, groupedBarCenter } from '../../core/bar-layout'
 import { Chart } from '../../core/Chart'
 import { ChartErrorBoundary } from '../../core/ChartErrorBoundary'
-import { barColorAt } from '../../core/color-utils'
 import { useLatest } from '../../core/hooks/useLatest'
 import {
     buildSegmentResolveValue,
@@ -47,17 +28,10 @@ import type {
     TooltipContext,
 } from '../../core/types'
 import { DEFAULT_Y_AXIS_ID } from '../../core/types'
-import { computeVisibleXLabels } from '../../overlays/AxisLabels'
 import { BarTooltip } from './BarTooltip'
-import { computeWrapperMinHeight, HORIZONTAL_MIN_BAND_SIZE_DEFAULT, resolveBarShadow } from './utils/bar-config'
-import {
-    barContainsPointOnBandAxis,
-    cursorOutsideBarFillExtent,
-    findVisibleStackedSegment,
-    groupedBandSlotAtCursor,
-    isStackedLayout,
-    iterBarsAtCursor,
-} from './utils/bars-under-cursor'
+import { computeWrapperMinHeight, HORIZONTAL_MIN_BAND_SIZE_DEFAULT } from './utils/bar-config'
+import { groupedBandSlotAtCursor } from './utils/bars-under-cursor'
+import { drawBarChartStatic, drawBarHoverItems, resolveBarHoverItems } from './utils/draw-bar-chart'
 import { resolveClickedBarSeries } from './utils/resolve-clicked-bar-series'
 
 export interface BarChartProps<Meta = unknown> {
@@ -72,37 +46,6 @@ export interface BarChartProps<Meta = unknown> {
     dataAttr?: string
     children?: React.ReactNode
     onError?: (error: Error, info: React.ErrorInfo) => void
-}
-
-const ALL_CORNERS: BarRoundedCorners = { topLeft: true, topRight: true, bottomLeft: true, bottomRight: true }
-
-/** One fully-rounded rect per band, spanning the union of that band's stacked segments — the
- *  pill the bar layer is clipped to for `roundStackEnds`. Bars in the same band share a band-axis
- *  slot (same `dataIndex`), so we group by it and extend along the value axis. */
-function stackPillRects(bars: BarRect[], isHorizontal: boolean): BarRect[] {
-    const byBand = new Map<number, BarRect>()
-    for (const bar of bars) {
-        if (bar.width <= 0 || bar.height <= 0) {
-            continue
-        }
-        const existing = byBand.get(bar.dataIndex)
-        if (!existing) {
-            byBand.set(bar.dataIndex, { ...bar, corners: ALL_CORNERS })
-            continue
-        }
-        if (isHorizontal) {
-            const left = Math.min(existing.x, bar.x)
-            const right = Math.max(existing.x + existing.width, bar.x + bar.width)
-            existing.x = left
-            existing.width = right - left
-        } else {
-            const top = Math.min(existing.y, bar.y)
-            const bottom = Math.max(existing.y + existing.height, bar.y + bar.height)
-            existing.y = top
-            existing.height = bottom - top
-        }
-    }
-    return [...byBand.values()]
 }
 
 export function BarChart<Meta = unknown>({ onError, ...rest }: BarChartProps<Meta>): React.ReactElement {
@@ -296,110 +239,20 @@ function BarChartInner<Meta = unknown>({
     )
 
     const drawStatic = useCallback(
-        ({ ctx, dimensions, scales, series: coloredSeries, labels: drawLabels, theme }: ChartDrawArgs) => {
-            const d3Scales = (scales._private as BarChartPrivate | undefined)?.__barChart
-            if (!d3Scales) {
-                return
-            }
-
-            const baseDrawCtx: DrawContext = {
-                ctx,
-                dimensions,
-                xScale: (label: string) => bandCenter(d3Scales, label),
-                yScale: d3Scales.value,
-                labels: drawLabels,
-            }
-
-            if (showGrid) {
-                // Align cross-axis grid with visible category labels, not every band.
-                let categoryTicks: number[] = []
-                if (isHorizontal) {
-                    for (const label of drawLabels) {
-                        const coord = bandCenter(d3Scales, label)
-                        if (coord != null && isFinite(coord)) {
-                            categoryTicks.push(coord)
-                        }
-                    }
-                } else {
-                    categoryTicks = computeVisibleXLabels(
-                        drawLabels,
-                        (label) => bandCenter(d3Scales, label),
-                        xTickFormatter
-                    ).map((entry) => entry.x)
-                }
-                drawGrid(baseDrawCtx, {
-                    gridColor: theme.gridColor,
-                    orientation: isHorizontal ? 'horizontal' : 'vertical',
-                    categoryTicks,
-                })
-            }
-
-            const seriesBars = coloredSeries
-                .filter((s) => !s.visibility?.excluded)
-                .map((s) => {
-                    const axisId = s.yAxisId ?? DEFAULT_Y_AXIS_ID
-                    const bars = computeSeriesBars({
-                        series: s,
-                        labels: drawLabels,
-                        scales: d3Scales,
-                        layout: barLayout,
-                        isHorizontal,
-                        stackedBand: stackedData?.get(s.key),
-                        isTopOfStack: topStackedKeyByAxis.get(axisId) === s.key,
-                    }).filter((b): b is BarRect => b !== null)
-                    return { series: s, bars }
-                })
-
-            // `roundStackEnds`: round both outer ends of the whole stack into a pill by clipping
-            // the bar layer to a rounded rect spanning each band's full extent, then drawing the
-            // segments square. The clip rounds the outer corners at the full radius even when the
-            // edge segment is a thin sliver (e.g. the last breakdown of a near-100% step), which
-            // per-segment rounding can't — it would clamp the radius to the sliver's half-width.
-            const stackPills = roundStackEnds
-                ? stackPillRects(
-                      seriesBars.flatMap((sb) => sb.bars),
-                      isHorizontal
-                  )
-                : []
-
-            // Tracks are a separate pass so a later series' full-height track can't paint
-            // over an earlier series' bar. Track is "share of a whole" semantics — only
-            // meaningful for grouped layouts; in stacked/percent every layer would paint
-            // a full-height track over the same band with the wrong corners.
-            if (barTrack && barLayout === 'grouped') {
-                const [axisStart = 0, axisEnd = 0] = d3Scales.value.range()
-                for (const { series: s, bars } of seriesBars) {
-                    const tracks = bars.map((b) => computeBarTrackRect(b, axisStart, axisEnd, isHorizontal))
-                    drawBarTracks(baseDrawCtx, s, tracks, barCornerRadius)
-                }
-            }
-
-            const resolvedShadow = resolveBarShadow(barShadow)
-            // Clip to plot area so a 100% bar's upward shadow doesn't bleed past the chart edge.
-            if (resolvedShadow) {
-                ctx.save()
-                ctx.beginPath()
-                ctx.rect(dimensions.plotLeft, dimensions.plotTop, dimensions.plotWidth, dimensions.plotHeight)
-                ctx.clip()
-                ctx.shadowColor = resolvedShadow.color
-                ctx.shadowBlur = resolvedShadow.blur
-                ctx.shadowOffsetX = resolvedShadow.offsetX ?? 0
-                ctx.shadowOffsetY = resolvedShadow.offsetY ?? 0
-            }
-            if (stackPills.length > 0) {
-                ctx.save()
-                clipToRoundedRects(ctx, stackPills, barCornerRadius)
-            }
-            for (const { series: s, bars } of seriesBars) {
-                drawBars(baseDrawCtx, s, bars, stackPills.length > 0 ? 0 : barCornerRadius, barFillStyle)
-            }
-            if (stackPills.length > 0) {
-                ctx.restore()
-            }
-            if (resolvedShadow) {
-                ctx.restore()
-            }
-        },
+        (args: ChartDrawArgs) =>
+            drawBarChartStatic(args, {
+                barLayout,
+                isHorizontal,
+                showGrid,
+                xTickFormatter,
+                stackedData,
+                topStackedKeyByAxis,
+                roundStackEnds,
+                barCornerRadius,
+                barTrack,
+                barShadow,
+                barFillStyle,
+            }),
         [
             showGrid,
             stackedData,
@@ -419,137 +272,34 @@ function BarChartInner<Meta = unknown>({
     const lastHoverKeyRef = useRef<string | null>(null)
 
     const drawHover = useCallback(
-        ({
-            ctx,
-            scales,
-            series: coloredSeries,
-            labels: drawLabels,
-            hoverIndex,
-            hoverPosition,
-            hoverProgress,
-            resetHoverFade,
-        }: ChartDrawArgs): DrawHoverResult => {
+        (args: ChartDrawArgs): DrawHoverResult => {
+            const { ctx, scales, hoverIndex, hoverProgress, resetHoverFade } = args
             const d3Scales = (scales._private as BarChartPrivate | undefined)?.__barChart
             if (!d3Scales || hoverIndex < 0) {
                 lastHoverKeyRef.current = null
                 return false
             }
-            const hoveredLabel = drawLabels[hoverIndex]
-            const [trackAxisStart = 0, trackAxisEnd = 0] = barTrack ? d3Scales.value.range() : []
-            // Key includes bar-vs-track per series so bar → track moves at the same
-            // hoverIndex still trigger a fade restart.
-            type DrawItem = { series: ResolvedSeries; bar: BarRect; isTrackHighlight: boolean }
-            const items: DrawItem[] = []
-            let composition = ''
-            // Stacked: clip the highlight to the visible slice so hover only changes shade,
-            // never z-order. Grouped keeps band-axis containment for cursor-above-bar.
-            const stackedHighlight = isStackedLayout(barLayout)
-            if (stackedHighlight && hoverPosition) {
-                const visible = findVisibleStackedSegment({
-                    series: coloredSeries,
-                    labels: drawLabels,
-                    hoveredLabel,
-                    cursor: hoverPosition,
-                    scales: d3Scales,
-                    layout: barLayout,
-                    isHorizontal,
-                    stackedData,
-                    topStackedKeyByAxis,
-                })
-                if (visible) {
-                    const visibleExtent = isHorizontal ? visible.bar.width : visible.bar.height
-                    const { nextSmallerExtent } = visible
-                    const baselinePx = isHorizontal ? visible.bar.x : visible.bar.y + visible.bar.height
-                    const clippedExtent = Math.max(0, visibleExtent - nextSmallerExtent)
-                    const clipped: BarRect = isHorizontal
-                        ? { ...visible.bar, x: baselinePx + nextSmallerExtent, width: clippedExtent }
-                        : { ...visible.bar, y: baselinePx - visibleExtent, height: clippedExtent }
-                    items.push({ series: visible.series, bar: clipped, isTrackHighlight: false })
-                    composition += 'b'
-                }
-            } else {
-                for (const { series: s, bar } of iterBarsAtCursor<ResolvedSeries>({
-                    series: coloredSeries,
-                    label: hoveredLabel,
-                    dataIndex: hoverIndex,
-                    scales: d3Scales,
-                    layout: barLayout,
-                    isHorizontal,
-                    stackedData,
-                    topStackedKeyByAxis,
-                })) {
-                    if (hoverPosition && !barContainsPointOnBandAxis(bar, hoverPosition, isHorizontal)) {
-                        continue
-                    }
-                    const isTrackHighlight =
-                        barTrack === true &&
-                        barLayout === 'grouped' &&
-                        hoverPosition != null &&
-                        cursorOutsideBarFillExtent(bar, hoverPosition, isHorizontal)
-                    items.push({ series: s, bar, isTrackHighlight })
-                    composition += isTrackHighlight ? 't' : 'b'
-                }
-            }
-            if (items.length === 0) {
+            const resolved = resolveBarHoverItems(args, d3Scales, {
+                barLayout,
+                isHorizontal,
+                stackedData,
+                topStackedKeyByAxis,
+                roundStackEnds,
+                barTrack,
+            })
+            if (!resolved) {
                 lastHoverKeyRef.current = null
                 return false
             }
-            const currentKey = `${hoverIndex}:${composition}`
+            // Key on the bar-vs-track composition so a bar → track move at the same hoverIndex
+            // still restarts the fade.
+            const currentKey = `${hoverIndex}:${resolved.composition}`
             let alpha = hoverProgress
             if (currentKey !== lastHoverKeyRef.current) {
                 alpha = resetHoverFade()
                 lastHoverKeyRef.current = currentKey
             }
-            // Match the resting bar's pill clip so the darker highlight rounds at the stack's outer
-            // ends instead of poking square corners past them.
-            const hoveredBandPills = roundStackEnds
-                ? stackPillRects(
-                      [
-                          ...iterBarsAtCursor<ResolvedSeries>({
-                              series: coloredSeries,
-                              label: hoveredLabel,
-                              dataIndex: hoverIndex,
-                              scales: d3Scales,
-                              layout: barLayout,
-                              isHorizontal,
-                              stackedData,
-                              topStackedKeyByAxis,
-                          }),
-                      ].map(({ bar }) => bar),
-                      isHorizontal
-                  )
-                : []
-            const highlightRadius = hoveredBandPills.length > 0 ? 0 : barCornerRadius
-            ctx.save()
-            ctx.globalAlpha = alpha
-            if (hoveredBandPills.length > 0) {
-                clipToRoundedRects(ctx, hoveredBandPills, barCornerRadius)
-            }
-            for (const { series: s, bar, isTrackHighlight } of items) {
-                if (isTrackHighlight) {
-                    const parsed = d3Color(barColorAt(s, bar.dataIndex))
-                    // Always translucent — the bar color direct would paint an opaque full-height
-                    // block if d3 can't parse the color.
-                    let trackColor: string
-                    if (parsed) {
-                        parsed.opacity = BAR_TRACK_HOVER_ALPHA
-                        trackColor = parsed.toString()
-                    } else {
-                        trackColor = `rgba(0,0,0,${BAR_TRACK_HOVER_ALPHA})`
-                    }
-                    drawBarHighlight(
-                        ctx,
-                        computeBarTrackRect(bar, trackAxisStart, trackAxisEnd, isHorizontal),
-                        trackColor,
-                        highlightRadius
-                    )
-                } else {
-                    const barColor = barColorAt(s, bar.dataIndex)
-                    const highlightColor = d3Color(barColor)?.darker(0.6).toString() ?? barColor
-                    drawBarHighlight(ctx, bar, highlightColor, highlightRadius)
-                }
-            }
-            ctx.restore()
+            drawBarHoverItems(ctx, d3Scales, resolved, { alpha, barCornerRadius, barTrack, isHorizontal })
             return true
         },
         [stackedData, barLayout, isHorizontal, topStackedKeyByAxis, roundStackEnds, barCornerRadius, barTrack]
